requirements.txt
fastapi
uvicorn
azure-ai-projects
azure-identity
python-dotenv
pydantic
requests
azure-monitor-opentelemetry